+++ src/induction/step.ml
@@ -722,6 +722,7 @@ let launch input_sys aparam trans =
   TransSys.define_and_declare_of_bounds
     trans
     (SMTSolver.define_fun solver)
+    ~define_rec:(SMTSolver.define_funs_rec solver)
     (SMTSolver.declare_fun solver)
     (SMTSolver.declare_sort solver)
     Numeral.zero Numeral.zero ;